+++ maplibre_gl_web/lib/src/options_sink.dart
@@ -46,4 +46,10 @@ abstract class MapLibreMapOptionsSink {
   void setScaleControlUnit(ScaleControlUnit unit);
 
   void setFeatureTapsTriggersMapClick(bool triggers);
+
+  void setLocationEngineProperties({
+    required bool enableHighAccuracy,
+    required int maximumAge,
+    required int timeout,
+  });
 }